backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
courlan==0.7.1
dateparser==1.1.1
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-multiselectfield==0.1.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
htmldate==1.2.1
huggingface-hub==0.5.1
idna==3.3
joblib==1.1.0
jusText==3.0.0
langcodes==3.3.0
lxml==4.8.0
numpy==1.22.3
packaging==21.3
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
sacremoses==0.0.49
six==1.16.0
sqlparse==0.4.2
tld==0.12.6
tokenizers==0.11.6
torch==1.11.0
tqdm==4.64.0
trafilatura==1.2.0
transformers==4.18.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wincertstore==0.2
